+++ bin/createSwooleTables.php
@@ -1,11 +1,14 @@
 <?php
 
+use Laravel\Octane\Tables\TableFactory;
 use Swoole\Table;
+
+require_once __DIR__.'/../src/Tables/TableFactory.php';
 
 $tables = [];
 
 foreach ($serverState['octaneConfig']['tables'] ?? [] as $name => $columns) {
-    $table = new Table(explode(':', $name)[1] ?? 1000);
+    $table = TableFactory::make(explode(':', $name)[1] ?? 1000);
 
     foreach ($columns ?? [] as $columnName => $column) {
         $table->column($columnName, match (explode(':', $column)[0] ?? 'string') {